admin deletion implemented

+++ src/routes/api/user/delete/+server.ts
@@ -2,8 +2,11 @@ import { prisma } from '$lib/server/prisma.server.js';
 import { delete_session } from '$lib/server/session.server';
 import { json } from '@sveltejs/kit';
 
-export async function POST({ cookies, locals }) {
+export async function POST({ cookies, request, locals }) {
     try {
+        // Get username of user to delete
+        const data = await request.json();
+        const username = data.username;
 
         if(!locals.user || !locals.authenticated) {
             return json(
@@ -12,11 +15,22 @@ export async function POST({ cookies, locals }) {
             );
         }
 
-        // delete user and corresponding data
+        const user = await prisma.users.findUnique({
+            where: { id: locals.user.id },
+            select: { rank: true }
+        });
+
+        // Validate user
+        if (locals.user.username !== username && user?.rank !== 'Admin') {
+            return json(
+                { error: 'Failed to delete user' },
+                { status: 401 }
+            );
+        }
+
+        // Delete user and corresponding data
         const result = await prisma.users.delete({
-            where: {
-                id: locals.user?.id
-            }
+            where: { username: username }
         });
 
         if(!result) {
@@ -26,19 +40,22 @@ export async function POST({ cookies, locals }) {
             );
         }
 
-        // Get the session ID from cookies
-        const session = cookies.get('session');
-    
-        if (session) {
-            // Delete the session
-            await delete_session(session);
-            cookies.delete('session', {path: '/',});
+        // If deleting own account, remove session
+        if (locals.user.username === username) {
+            // Get the session ID from cookies
+            const session = cookies.get('session');
+        
+            if (session) {
+                // Delete the session
+                await delete_session(session);
+                cookies.delete('session', {path: '/',});
+            }
+            
+            locals.user = null;
+            locals.authenticated = false;
         }
         
-        locals.user = null;
-        locals.authenticated = false;
         return json({ success: true });
-    
     } catch (e) {
         return json({ error: `Internal server error: ${e}` }, { status: 500 });
   }
